extracted/manus-manager-updated/manus-manager/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.6
websockets==11.0.3
pytest==7.3.1
httpx==0.24.0
alembic==1.10.4
python-dotenv==1.0.0
requests==2.29.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
